+++ programs-ecs/bundles/ecs-bundle/src/systems/smart_object_deity_interact/mod.rs
@@ -1,4 +1,4 @@
-use bolt_lang::error_code;
+use bolt_lang::*;
 
 #[error_code]
 pub enum SmartObjectDeityInteractionError {
