+++ programs-ecs/bundles/ecs-bundle/src/systems/smart_object_token_launcher_init/mod.rs
@@ -1,4 +1,4 @@
-use bolt_lang::error_code;
+use bolt_lang::*;
 
 #[error_code]
 pub enum SmartObjectTokenLauncherInitError {
