--- programs-ecs/systems/smart-object-deity-interact/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "smart-object-deity-interact"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "smart_object_deity_interact"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build",  "anchor-spl/idl-build", "ecs-bundle/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
-deity-bot = { version = "0.1.0", path = "../../../programs/deity-bot", features = ["cpi"] }
-anchor-spl.workspace = true
