--- programs-ecs/systems/smart-object-token-launcher-init/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "smart-object-token-launcher-init"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "smart_object_token_launcher_init"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build", "ecs-bundle/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-serde.workspace = true
-anchor-spl.workspace = true
-ecs-bundle.workspace = true
