--- programs-ecs/systems/smart-object-token-launcher-init/Xargo.toml
@@ -1,2 +1,0 @@
-[target.bpfel-unknown-unknown.dependencies.std]
-features = []
