--- programs-ecs/systems/smart-object-token-launcher-interact/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "smart-object-token-launcher-interact"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "smart_object_token_launcher_interact"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build", "ecs-bundle/idl-build"]
-anchor-debug = []
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
-anchor-spl.workspace = true
